Don't assume PWD is the checkout root.
The AppImage version of love2d changes to a temporary directory before
running, so you have to explicitly specify the full path; relative
paths won't be correct.

This won't fix the problem when running from a .love file; in that
case you would need to do AOT compilation and include the .lua output
inside the archive, but it's a start for folks running from source.